Make some OS X package upgrades, and make some Makefile fixes in order to allow Armory to compile post-megamerge. Armory still won't run under OS X and will require more Makefile fixes.

+++ cppForSwig/cryptopp/wake.cpp
@@ -22,16 +22,20 @@ void WAKE_Base::GenKey(word32 k0, word32 k1, word32 k2, word32 k3)
 	// this code is mostly copied from David Wheeler's paper "A Bulk Data Encryption Algorithm"
 	signed int x, z, p;	
 	// x and z were declared as "long" in Wheeler's paper, which is a signed type. I don't know if that was intentional, but it's too late to change it now. -- Wei 7/4/2010
+    // ARMORY - In order to compile for C++11, some values get the
+    // static_cast treatment. Will this screw up WAKE
+    // (https://en.wikipedia.org/wiki/WAKE_%28cipher%29)? Who knows. Armory
+    // doesn't use WAKE, so we don't care if it breaks. :)
 	CRYPTOPP_COMPILE_ASSERT(sizeof(x) == 4);
 	static int tt[10]= {
 		0x726a8f3b,								 // table
-		0xe69a3b5c,
-		0xd3c71fe5,
-		0xab3c73d2,
+		static_cast<int>(0xe69a3b5c),
+		static_cast<int>(0xd3c71fe5),
+		static_cast<int>(0xab3c73d2),
 		0x4d3a8eb3,
 		0x0396d6e8,
 		0x3d4c2f7a,
-		0x9ee27cf3, } ;
+		static_cast<int>(0x9ee27cf3) } ;
 	t[0] = k0;
 	t[1] = k1;
 	t[2] = k2;
